Fix: 위도 경도 저장 타입 변경

See merge request s08-final/S08P31C101!691

+++ backend/product-service/src/main/java/com/example/productservice/dto/request/ReservationReturnRequestDto.java
@@ -13,9 +13,9 @@ public class ReservationReturnRequestDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
     private LocalDate endDate;
 
-    private Float lng;
+    private Double lng;
 
-    private Float lat;
+    private Double lat;
 
     private String location;
 }
